+++ doc/whats_new.rst
@@ -131,6 +131,9 @@ Changelog
      :mod:`sklearn.metrics` for regression and classification metrics
      by `Arnaud Joly`_.
 
+   - Fixed a bug in :class:`sklearn.svm.SVC` when using csr-matrices with
+     unsorted indices by Meng Xinfan and `Andreas Müller`_.
+
 API changes summary
 -------------------
    - Renamed all occurences of ``n_atoms`` to ``n_components`` for consistency.